+++ src/components/GettingStarted/GetStarted.jsx
