+++ src/index.js
@@ -4,7 +4,7 @@ import ReactDOM from "react-dom";
 import "./index.css";
 import App from "./App";
 import { Provider } from "react-redux";
-import store from "./Redux/store";
+import store from "./redux/store";
 
 ReactDOM.render(
   <React.StrictMode>
